+++ asn1-tools/enber/Makefile.am
@@ -1,0 +1,9 @@
+@CODE_COVERAGE_RULES@
+
+AM_CFLAGS = @ADD_CFLAGS@
+AM_CPPFLAGS =                       \
+    -I${top_srcdir}/libasn1common   \
+    -I${top_srcdir}/libasn1parser   \
+    -I${top_srcdir}/skeletons
+
+bin_PROGRAMS = enber
