+++ asn1c/Makefile.am
@@ -17,14 +17,4 @@ asn1c_LDADD =                                           \
     $(top_builddir)/libasn1fix/libasn1fix.la            \
     $(top_builddir)/libasn1compiler/libasn1compiler.la
 
-unber_LDADD =                                           \
-    $(top_builddir)/libasn1common/libasn1common.la
-
-bin_PROGRAMS = asn1c unber enber
-
-noinst_HEADERS = sys-common.h
-
-dist_check_SCRIPTS = check-xxber.sh
-TESTS_ENVIRONMENT= top_srcdir=${top_srcdir}
-TESTS = $(dist_check_SCRIPTS)
-CLEANFILES = .check-xxber.*.tmp
+bin_PROGRAMS = asn1c
