+++ libasn1common/Makefile.am
@@ -5,7 +5,7 @@ AM_CFLAGS = @ADD_CFLAGS@
 noinst_LTLIBRARIES = libasn1common.la
 
 libasn1common_la_SOURCES =              \
-    asn1_ref.c asn1_ref.h               \
+    asn1_common.h asn1_ref.c asn1_ref.h \
     asn1_buffer.c asn1_buffer.h         \
     asn1_namespace.c asn1_namespace.h   \
     genhash.c genhash.h
